2010-05-27	Ralf Corsépius <ralf.corsepius@rtems.org>
	* libcsupport/include/rtems/libio_.h, libcsupport/src/eval.c:
	Use size_t instead of int for pathnamelen.

+++ cpukit/libcsupport/include/rtems/libio_.h
@@ -246,7 +246,7 @@ int rtems_libio_is_file_open(
 
 int rtems_filesystem_evaluate_relative_path(
   const char                        *pathname,
-  int                                pathnamelen,
+  size_t                             pathnamelen,
   int                                flags,
   rtems_filesystem_location_info_t  *pathloc,
   int                                follow_link
@@ -254,7 +254,7 @@ int rtems_filesystem_evaluate_relative_path(
 
 int rtems_filesystem_evaluate_path(
   const char                        *pathname,
-  int                                pathnamelen,
+  size_t                             pathnamelen,
   int                                flags,
   rtems_filesystem_location_info_t  *pathloc,
   int                                follow_link
